implement authentication on posts endpoints

+++ src/main/java/com/github/terravivaproject/terraviva/social/services/PostService.java
@@ -1,16 +1,19 @@
 package com.github.terravivaproject.terraviva.social.services;
 
+import com.github.terravivaproject.terraviva.exceptions.EntityDoesNotExist;
+import com.github.terravivaproject.terraviva.exceptions.UnauthorizedException;
+import com.github.terravivaproject.terraviva.resources.ErrorMessagesService;
 import com.github.terravivaproject.terraviva.social.entities.Post;
 import com.github.terravivaproject.terraviva.social.entities.Tag;
-import com.github.terravivaproject.terraviva.social.entities.dto.CreationPostDto;
-import com.github.terravivaproject.terraviva.social.entities.dto.PostDto;
+import com.github.terravivaproject.terraviva.social.entities.dto.PostRto;
+import com.github.terravivaproject.terraviva.social.entities.mappers.PostMapper;
 import com.github.terravivaproject.terraviva.social.repositories.PostRepository;
 import com.github.terravivaproject.terraviva.user.entities.AppUser;
-import com.github.terravivaproject.terraviva.user.repositories.UserRepository;
+import com.github.terravivaproject.terraviva.user.services.UserService;
+import dev.dmgiangi.budssecurity.securitycontext.SecurityContext;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 
@@ -18,45 +21,51 @@ import java.util.UUID;
 @AllArgsConstructor
 public class PostService {
 
-    private PostRepository postRepository;
+    private final PostRepository postRepository;
+    private final TagService tagService;
+    private final UserService userService;
 
-    private UserRepository userRepository;
+    public Post createPost(PostRto postRequest) {
+        //Map a set Of Tag entities from the strings found in PostRto
+        Set<Tag> tags = tagService.tagFromStrings(postRequest.getTags());
 
-    private TagService tagService;
+        //Get the user from the UserService using the Security user from the SecurityContext
+        AppUser postOwner = userService.getUserById(
+                UUID.fromString(
+                        SecurityContext.getUser().getMainIdentifier()
+                )
+        );
 
-    public PostDto createPost(CreationPostDto creationPostDto) {
-        Optional<AppUser> owner = userRepository.findById(creationPostDto.getOwner());
-        if (owner.isEmpty()) throw new RuntimeException("The user does not exist");
-        Set<Tag> tags = tagService.tagFromStrings(creationPostDto.getTags());
+        //create a new post entity
+        Post newPost = PostMapper.MAP.rtoToEntity(postRequest, postOwner, tags);
 
-        Post post = new Post();
-        post.setTags(tags)
-                .setOwner(owner.get())
-                .setMessage(creationPostDto.getMessage());
-
-        postRepository.save(post);
-
-        return new PostDto()
-                .setId(post.getId())
-                .setMessage(post.getMessage())
-                .setOwner(post.getOwner().getId())
-                .setTags(post.getTags().stream().map(Tag::getName).toList())
-                .setDateTime(post.getDateTime());
+        return postRepository.save(newPost);
     }
 
-    public PostDto getPostId(UUID id) {
-        Optional<Post> post = postRepository.findById(id);
-        if (post.isEmpty()) throw new RuntimeException("This post does not exist");
-        Set<Tag> tags = post.get().getTags();
-        return new PostDto()
-                .setDateTime(post.get().getDateTime())
-                .setMessage(post.get().getMessage())
-                .setId(post.get().getId())
-                .setOwner(post.get().getOwner().getId())
-                .setTags(tags.stream().map(Tag::getName).toList());
+    public Post getPostId(UUID id) {
+        return postRepository.findById(id)
+                .orElseThrow(() -> new EntityDoesNotExist(
+                        ErrorMessagesService.resourceNotExist()
+                ));
     }
 
     public void deletePost(UUID id) {
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new EntityDoesNotExist(
+                        ErrorMessagesService.resourceNotExist()
+                ));
+
+        UUID authenticatedUserId =
+                UUID.fromString(
+                        SecurityContext.getUser().getMainIdentifier()
+                );
+
+        // TODO: 23/09/22 permits roles check
+        if (!post.getOwner().getId().equals(authenticatedUserId))
+            throw new UnauthorizedException(
+                    ErrorMessagesService.notEnoughAuthorization()
+            );
+
         postRepository.deleteById(id);
     }
 }
